refactor: clean up auth module and related files, remove unused code and guards
