--- apps/hr-core-server/src/modules/auth/strategies/jwt.constants.strategy.ts
@@ -1,5 +1,0 @@
-export const jwtConstants = {
-  secret: process.env.JWT_SECRET || 'defaultSecretKey',
-  expiresIn: '1d', // access token
-  refreshIn: '7d', // refresh token
-};
