+++ libs/hr/core-app/src/index.ts
@@ -1,4 +1,5 @@
 export * from './lib/core-app.module';
-export * from './lib/auth/auth.module';
+// export * from './lib/auth/auth.module';
 export * from './lib/core/common.module';
 export * from './lib/core/infrastructure/midilwares/index';
+export * from './lib/auth/index';
